Feature/calendar

+++ src/main/java/kr/carz/savecar/dto/MonthlyRentVO.java
@@ -17,10 +17,10 @@ public class MonthlyRentVO {
     private String name;
 
     private String deposit;
-    private Float cost_for_2k;
-    private Float cost_for_2_5k;
-    private Float cost_for_3k;
-    private Float cost_for_4k;
+    private double cost_for_2k;
+    private double cost_for_2_5k;
+    private double cost_for_3k;
+    private double cost_for_4k;
     private String cost_for_others;
     private String age_limit;
     private String cost_per_km;
